feat: 加入音频系统，音频服务层 feat: 事件系统加入立即事件，事件类型加入音频事件，加入状态操作请求事件 feat: Rect加入hasIntersection判定 feat: 对象池加入自定义删除器，通过conditional自动判断 feat: 资源管理器加入音频读取 feat: 加入GameSession系统 feat: mapLoader加入对象读取功能 feat: 加入玩家和敌人血条显示 feat: 加入新状态选择人物界面 feat: 添加地图对象层 refactor: 敌人AI寻路机制完善 refactor: 状态机加入状态栈 refactor: 所有更新接口参数都修改为double双精度浮点数 refactor: 更改输入系统鼠标更新方式

+++ Platform jumping game 2D _based on sdl3/EnemyManager.cpp
@@ -5,7 +5,7 @@ EnemyManager::EnemyManager(Renderer& renderer, Resource& rM, size_t poolSize) no
 	
 }
 
-void EnemyManager::update(float dt) noexcept {
+void EnemyManager::update(double dt) noexcept {
 	for (Enemy* enemy : activeEnemies_) {
 		enemy->update(dt);
 	}
