+++ Platform jumping game 2D _based on sdl3/WonState.cpp
@@ -11,6 +11,6 @@ void WonState::render() const noexcept{
 	renderer_.clearDynamicTexture(); // 切换到动态纹理渲染目标并清除内容，准备渲染胜利界面等动态元素
 }
 
-void WonState::update(float dt) noexcept{
+void WonState::update(double dt) noexcept{
 
 }
